testpmd: Move to a new file
This commit moves code to a new file without changing it.
The new file creates a separation between the testpmd related
and the rest of the executor code, preparing it to be move to
another package in future commits.

Also Adding licence paragraph.
